+++ enumeration/remote/aws/init.go
@@ -1,6 +1,7 @@
 package aws
 
 import (
+	"github.com/aws/aws-sdk-go-v2/aws"
 	"github.com/snyk/driftctl/enumeration"
 	"github.com/snyk/driftctl/enumeration/alerter"
 	"github.com/snyk/driftctl/enumeration/remote/aws/repository"
@@ -16,19 +17,20 @@ import (
  */
 
 // Init configures the AWS provider and registers its enumerators with the remote library.
-func Init(version string, _ alerter.Interface, providerLibrary *terraform.ProviderLibrary, remoteLibrary *common.RemoteLibrary, progress enumeration.ProgressCounter, factory resource.Factory, configDir string) error {
+// It returns the aws.Config so callers can create additional AWS service clients.
+func Init(version string, _ alerter.Interface, providerLibrary *terraform.ProviderLibrary, remoteLibrary *common.RemoteLibrary, progress enumeration.ProgressCounter, factory resource.Factory, configDir string) (aws.Config, error) {
 
 	provider, err := NewTerraformProvider(version, progress, configDir)
 	if err != nil {
-		return err
+		return aws.Config{}, err
 	}
 	err = provider.CheckCredentialsExist()
 	if err != nil {
-		return err
+		return aws.Config{}, err
 	}
 	err = provider.Init()
 	if err != nil {
-		return err
+		return aws.Config{}, err
 	}
 
 	repositoryCache := cache.New(100)
@@ -39,5 +41,5 @@ func Init(version string, _ alerter.Interface, providerLibrary *terraform.Provid
 	configEnumerator := NewConfigEnumerator(configRepo, factory)
 	remoteLibrary.AddBulkEnumerator(configEnumerator)
 
-	return nil
+	return provider.AwsCfg, nil
 }